ok realigning tabs looks good DO NOT FUCKING CHANGE IT AGAIN
as far as im aware the only thing that needs fixing now is the selection of tabs in the unassigned groups list. possibly branching into more of a correctly identifying matching tab infos as when I close a tab and reopen it it gets a bit confused. possible that create new tab section is the culprit