Fixed Selection Stuff
There is an underlying issue of saving and loading, going to need to fix that next. I thin kthe best course of action is to store the tab infos and tab groups separately and do away with the TabsInGroup part of the tab groups, as we can just realign them later and it would save me having to loop through each individual group to add a unique tab group to the list.